+++ packages-js/ui/src/rest/index.ts
@@ -1,3 +1,3 @@
-export { RestError, requestFile, requestJson, type DownloadFile, type FetchLike, type RestOptions } from './client'
+export { RestError, requestFile, requestJson, requestUpload, type DownloadFile, type FetchLike, type RestOptions } from './client'
 export { saveFile } from './download'
 export { createRunner, type Runner } from './runner'